Move from duckdb/wrapped/std to duckdb/original/std, and review comments
