+++ src/include/duckdb/original/std/locale.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <cctype>
+#include <locale>
+
+#define DUCKDB_BASE_STD
+
+namespace duckdb_base_std {
+using ::std::isspace;
+} // namespace duckdb_base_std
