+++ src/include/duckdb/original/std/memory.hpp
@@ -2,13 +2,11 @@
 
 #include <memory>
 
-#define DUCKDB_WRAP_STD
+#define DUCKDB_BASE_STD
 
-namespace duckdb_wrapped {
-namespace std {
+namespace duckdb_base_std {
 using ::std::make_shared;
 using ::std::shared_ptr;
 using ::std::unique_ptr;
 // using ::std::make_unique;
-} // namespace std
-} // namespace duckdb_wrapped
+} // namespace duckdb_base_std
