+++ src/include/duckdb/original/std/sstream.hpp
@@ -2,12 +2,10 @@
 
 #include <sstream>
 
-#define DUCKDB_WRAP_STD
+#define DUCKDB_BASE_STD
 
-namespace duckdb_wrapped {
-namespace std {
+namespace duckdb_base_std {
 using ::std::basic_stringstream;
 using ::std::stringstream;
 using ::std::wstringstream;
-} // namespace std
-} // namespace duckdb_wrapped
+} // namespace duckdb_base_std
