--- src/include/duckdb/wrapped/std/locale.hpp
@@ -1,12 +1,0 @@
-#pragma once
-
-#include <cctype>
-#include <locale>
-
-#define DUCKDB_WRAP_STD
-
-namespace duckdb_wrapped {
-namespace std {
-using ::std::isspace;
-} // namespace std
-} // namespace duckdb_wrapped
